デフォルトの状態で連続実行しないよう変更
